Allow tests to pass after upgrade to .net core
I updated the packages and noticed that with the XUnit upgrade, the more
idiomatic way of handling async setup is by implementing IAsyncLifetime, so I
did this.

I also needed to update the status code slightly on duration - possibly the mpd
version has changed. I am testing with 0.21.20, which returns duration in a
"xxx.yyy" format.